requirements.txt
openai==0.28.0
PyMuPDF==1.20.1  
Flask==2.2.2
Werkzeug==2.2.2
python-docx==0.8.11